steps towards being able to do a slick user dao
Former-commit-id: bae3249cd532893da9dd06ccc204193239610515
Former-commit-id: 57232e0a42dfc8cd95057362d960c8eef7c1c062

+++ src/mitll/langtest/shared/User.java
@@ -5,6 +5,7 @@
 package mitll.langtest.shared;
 
 import com.google.gwt.user.client.rpc.IsSerializable;
+import mitll.langtest.server.database.user.UserDAO;
 
 import java.util.Collection;
 
@@ -60,14 +61,14 @@ public class User extends MiniUser {
    * @param enabled
    * @param permissions
    */
-  public User(long id, int age, int gender, int experience, String ipaddr, String password,
+  public User(int id, int age, int gender, int experience, String ipaddr, String password,
               boolean enabled, Collection<Permission> permissions) {
     this(id, age, gender, experience, ipaddr, password, NOT_SET, NOT_SET, NOT_SET, enabled, false, permissions,
         Kind.STUDENT, "", "", "", "", System.currentTimeMillis());
   }
 
   /**
-   * @see mitll.langtest.server.database.UserDAO#getUsers()
+   * @see UserDAO#getUsers()
    * @param id
    * @param age
    * @param gender
@@ -85,7 +86,7 @@ public class User extends MiniUser {
    * @param cdEnableKey
    * @param timestamp
    */
-  public User(long id, int age, int gender, int experience, String ipaddr, String passwordH,
+  public User(int id, int age, int gender, int experience, String ipaddr, String passwordH,
               String nativeLang, String dialect, String userID, boolean enabled, boolean isAdmin,
               Collection<Permission> permissions, Kind userKind, String emailHash, String device, String resetPassKey,
               String cdEnableKey, long timestamp) {
