+++ src/mitll/langtest/shared/UserAndTime.java
@@ -8,7 +8,7 @@ package mitll.langtest.shared;
  * Created by go22670 on 9/17/15.
  */
 public interface UserAndTime {
-  long getUserid();
+  int getUserid();
   long getTimestamp();
   String getExid();
   String getID();
